+++ R/eng-mod_ts_ita.R
@@ -10,11 +10,14 @@
 #' @importFrom plotly plotlyOutput
 eng_mod_ts_ita_ui <- function(id, title, width = 12){
   ns <- NS(id)
+  fluidPage(
+    fluidRow(shiny::checkboxInput(ns("y_log"), "Scala logaritmica")),
   fluidRow(
     box(title = title, width = width,
       plotlyOutput(ns("ts_plot"))
     )
   )
+ )
 }
 
 #' ts_ita Server Function
@@ -78,7 +81,18 @@ eng_mod_ts_ita_server <- function(id, type = c("cum", "inc")) {
 
   callModule(id = id, function(input, output, session) {
     ns <- session$ns
-    output$ts_plot <- renderPlotly(ggplotly(gg))
+    output$ts_plot <- renderPlotly({
+      if (input$y_log) {
+        gg <- gg + scale_y_continuous(
+          trans = 'log2',
+          breaks = scales::trans_breaks("log2", function(x) 2^x),
+          labels = scales::trans_format("log2", scales::math_format(2^.x))
+        ) +
+          ylab(paste0(y_lab," - log2"))
+      }
+
+      ggplotly(gg)
+    })
   })
 }