+++ NEWS.md
@@ -1,3 +1,4 @@
+* logarithmic scale - eng (#5)
 * Solve bug (#31) 
 * Added Venezia to default provinces (fix #30)